--- a/cl-quil.asd
+++ b/cl-quil.asd
@@ -8,14 +8,12 @@
   :license "Apache License 2.0 (See LICENSE.txt)"
   :version (:read-file-form "VERSION.txt")
   :pathname "src/"
-  :depends-on ((:version #:alexa "2.1.1")
-                                        ; Lexical analysis
-               #:yacc                   ; Arithmetic parsing
-               #:alexandria             ; Utilities
-               #:parse-float            ; Float parsing
+  :depends-on (#:cl-quil/frontend
+               #:alexandria
                (:version #:magicl/core "0.10.0")
                #:magicl/ext-lapack      ; for CSD
                                         ; internal linear algebra library
+               #:qvm
                #:cl-grnm                ; nelder-mead implementation
                #:singleton-classes
                #:abstract-classes
@@ -25,57 +23,27 @@
                #:closer-mop
                #:optima
                #:cl-algebraic-data-type
-               #:global-vars            ; Static globals
-               #:salza2                 ; God table compression
+               #:global-vars
                #:trivial-garbage        ; weak hash tables
                #:flexi-streams          ; For executable writing
                #:cl-heap
                #:cl-permutation
                #:queues.priority-queue
-               #+sbcl #:sb-rotate-byte
                )
   :in-order-to ((asdf:test-op (asdf:test-op #:cl-quil-tests)))
   :around-compile (lambda (compile)
                     (let (#+sbcl (sb-ext:*derive-function-types* t))
                       (funcall compile)))
   :serial t
-  :components ((:module "quil"
-                :serial t
-                :components ((:static-file "stdgates.quil")))
-               (:file "package")
+  :components ((:file "package")
                (:file "options")
-               (:file "types")
                (:file "utilities")
-               (:file "queue")
-               (:file "magicl-constructors")
                (:file "relabeling")
                (:file "matrix-operations")
                (:file "csd")
-               (:file "transformable-mixin")
-               (:file "classical-memory")
-               (:file "ast")
                (:file "resource")
-               (:file "define-pragma")
-               (:file "pragmas")
-               (:file "parser")
-               (:file "cl-quil")
-               (:file "qasm")
-               (:file "gates")
                (:file "context")
-               (:file "build-gate")
                (:file "define-compiler")
-               (:module "clifford"
-                :serial t
-                :components ((:file "qubit-algebra")
-                             (:file "pauli")
-                             (:file "symplectic")
-                             (:file "clifford")
-                             (:file "stabilizer")
-                             (:file "god-table")
-                             (:file "god-table-utilities")
-                             (:file "swap-representation")
-                             (:file "benchmarking-procedures")
-                             (:file "perm")))
                (:module "compilers"
                 :serial t
                 :components ((:file "ucr-explode")
@@ -94,19 +62,6 @@
                              ;; attic'd files / pedagogical purposes only
                              (:static-file "optimal-2q")
                              (:static-file "cs-compile")))
-               (:module "analysis"
-                :serial t
-                :components ((:file "process-includes")
-                             (:file "type-safety")
-                             (:file "patch-labels")
-                             (:file "resolve-objects")
-                             (:file "qubits-needed")
-                             (:file "compress-qubits")
-                             (:file "expansion")
-                             (:file "expand-circuits")
-                             (:file "rewrite-arithmetic")
-                             (:file "fusion")
-                             (:file "simplify-arithmetic")))
                (:file "cfg")
                (:file "compilation-methods")
                (:module "chip"
@@ -141,6 +96,76 @@
                              (:file "wavefunctions")
                              (:file "rewriting-rules")))))
 
+(asdf:defsystem #:cl-quil/frontend
+  :description "Syntax, parsers, and basic analysis routines for the Quil language."
+  :version (:read-file-form "VERSION.txt")
+  :depends-on ((:version #:alexa "2.1.1")
+                                        ; Lexical analysis
+               #:yacc                   ; Arithmetic parsing
+               #:alexandria             ; Utilities
+               #:parse-float            ; Float parsing
+               #:singleton-classes
+               #:abstract-classes
+               #:split-sequence
+               #:cl-algebraic-data-type
+               #:cl-permutation
+               #:trivial-garbage
+               #:magicl/core
+               #:magicl/ext-lapack      ; for gate fusion
+               #:global-vars            ; Static globals
+               #:salza2                 ; God table compression
+               #:optima
+
+               #+sbcl #:sb-rotate-byte
+               )
+  :pathname "src/"
+  :serial t
+  :components ((:module "quil"
+                :serial t
+                :components ((:static-file "stdgates.quil")))
+               (:file "package")
+               (:file "frontend-options")
+               (:file "types")
+               (:file "frontend-utilities")
+               (:file "queue")
+               (:file "magicl-constructors")
+               (:file "classical-memory")
+               (:file "ast")
+               (:file "define-pragma")
+               (:file "pragmas")
+               (:file "parser")
+               (:file "cl-quil")
+               (:file "qasm")
+               (:file "gates")
+               (:file "build-gate")
+               (:file "with-inst")
+               (:file "transformable-mixin") 
+               (:module "clifford"
+                :serial t
+                :components ((:file "qubit-algebra")
+                             (:file "pauli")
+                             (:file "symplectic")
+                             (:file "clifford")
+                             (:file "stabilizer")
+                             (:file "god-table")
+                             (:file "god-table-utilities")
+                             (:file "swap-representation")
+                             (:file "benchmarking-procedures")
+                             (:file "perm")))
+               (:module "analysis"
+                :serial t
+                :components ((:file "process-includes")
+                             (:file "type-safety")
+                             (:file "patch-labels")
+                             (:file "resolve-objects")
+                             (:file "qubits-needed")
+                             (:file "compress-qubits")
+                             (:file "expansion")
+                             (:file "expand-circuits")
+                             (:file "rewrite-arithmetic")
+                             (:file "fusion")
+                             (:file "simplify-arithmetic")))))
+
 ;;; Contribs
 
 ;; Adapted from magicl's MAGICL/EXT-EXPOKIT adapted from commonqt's
@@ -151,7 +176,7 @@
   :license "Apache License 2.0 (See LICENSE.txt)"
   :maintainer "Rigetti Computing"
   :author "Rigetti Computing"
-  :depends-on (#:cl-quil)
+  :depends-on (#:cl-quil/frontend)
   :in-order-to ((asdf:test-op (asdf:test-op #:cl-quil/quilt-tests)))
   :around-compile (lambda (compile)
                     (let (#+sbcl (sb-ext:*derive-function-types* t))