--- a/src/addresser/rewiring.lisp
+++ b/src/addresser/rewiring.lisp
@@ -2,21 +2,7 @@
 ;;;;
 ;;;; Author: Eric Peterson, Corwin de Boor
 
-;; We store both the l2p and p2l vectors so that lookups in both
-;; directions can be constant time. Because all of our qubits are in
-;; the range 0...n-1, we can store these as vectors instead of hashmaps.
-
 (in-package #:cl-quil)
-
-(defstruct (rewiring
-            (:constructor init-rewiring)
-            (:copier nil))
-  (l2p #() :type integeropt-vector)
-  (p2l #() :type integeropt-vector))
-
-(defmethod print-object :around ((object rewiring) stream)
-  (let ((*print-pretty* nil))
-    (call-next-method)))
 
 (defun inverse-matches-forward-p (forward inverse)
   "Checks that each non-NIL mapping in FORWARD has a corresponding inverse mapping in
@@ -214,19 +200,6 @@ Returns NIL. This mutates the instruction."
   ;; Return nil to emphasize side effect.
   nil)
 
-(defun make-rewiring-from-l2p (l2p)
-  "Safely extract a REWIRING from a logical-to-physical array."
-  (let ((p2l (make-array (length l2p) :initial-element nil)))
-    (dotimes (j (length l2p))
-      (let ((loc (aref l2p j)))
-        (when loc
-          (assert (and (< -1 loc (length p2l))) ()
-                  "Malformed rewiring string: value ~A at position ~A is out of range." loc j)
-          (assert (null (aref p2l loc)) ()
-                  "Malformed rewiring string: repeated value ~A at position ~A." loc j)
-          (setf (aref p2l loc) j))))
-    (init-rewiring :l2p l2p :p2l p2l)))
-
 (defmacro with-update-rewiring (rewiring q0 q1 &body body)
   "Temporarily assigns the LOGICAL to PHYSICAL in REWIRING, and then executes
 BODY as an implicit PROGN."
@@ -252,66 +225,6 @@ BODY as an implicit PROGN."
        (rewiring-assign ,rewiring-sym ,logical-sym ,physical-sym)
        (unwind-protect (progn ,@body)
          (rewiring-unassign ,rewiring-sym ,logical-sym)))))
-
-(defun make-rewiring-from-string (str)
-  "Safely extract a REWIRING from a string representation of an integer vector corresponding to the logical-to-physical mapping."
-  (assert (and (eql #\# (aref str 0))
-               (eql #\( (aref str 1))
-               (eql #\) (aref str (1- (length str)))))
-          nil
-          "Malformed rewiring string: input ~A is not of the form #(...)." str)
-  (let* ((stripped-string (string-trim "#()" str))
-         (tokens (first (tokenize stripped-string)))
-         (integer-vec
-           (map 'vector
-                (lambda (token)
-                  (cond
-                    ((equalp (token-payload token) "nil")
-                     nil)
-                    ((eql (token-type token) :integer)
-                     (token-payload token))
-                    (t
-                     (error "Malformed rewiring string: unexpected token ~A." token))))
-                tokens)))
-    (make-rewiring-from-l2p integer-vec)))
-
-(defun make-rewiring-pair-from-string (str)
-  "Safely extract a pair of REWIRINGs from a string representation of a CONS of two integer vectors."
-  (multiple-value-bind (matchedp matches)
-      ;; This monstrosity matches strings of the form "(#(\d+ ...) . #(\d+ ...))"
-      (let ((match-int-vector
-              '(:REGISTER
-                (:SEQUENCE "#("
-                 (:GREEDY-REPETITION 0 NIL
-                  (:GROUP (:SEQUENCE :DIGIT-CLASS (:GREEDY-REPETITION 0 1 #\ ))))
-                 #\)))))
-        (cl-ppcre:scan-to-strings
-         `(:SEQUENCE
-           :START-ANCHOR
-           #\(
-           ,match-int-vector
-           (:GREEDY-REPETITION 0 NIL :WHITESPACE-CHAR-CLASS)
-           #\.
-           (:GREEDY-REPETITION 0 NIL :WHITESPACE-CHAR-CLASS)
-           ,match-int-vector
-           #\)
-           :END-ANCHOR)
-         str))
-    (assert matchedp
-            nil
-            "Malformed rewiring pair string: ~@
-             input ~A is not of the form (#(...) . #(...))."
-            str)
-    (let ((first-rewiring-string (aref matches 0))
-          (second-rewiring-string (aref matches 1)))
-      (assert (= (length first-rewiring-string) (length second-rewiring-string))
-              nil
-              "Malformed rewiring pair string: length of rewirings don't match. ~@
-               first:  ~A~@
-               second: ~A"
-              first-rewiring-string second-rewiring-string)
-      (values (make-rewiring-from-string first-rewiring-string)
-              (make-rewiring-from-string second-rewiring-string)))))
 
 (defun rewiring-assigned-for-qubit-p (rewiring lq)
   "Test whether REWIRING contains a non-nil rewiring for logical qubit LQ."