--- a/src/analysis/compress-qubits.lisp
+++ b/src/analysis/compress-qubits.lisp
@@ -2,7 +2,7 @@
 ;;;;
 ;;;; Author: Robert Smith
 
-(in-package #:cl-quil)
+(in-package #:cl-quil.frontend)
 
 ;;; This file contains an analysis to compute the qubits used by a
 ;;; program.