--- a/src/analysis/expand-circuits.lisp
+++ b/src/analysis/expand-circuits.lisp
@@ -2,7 +2,7 @@
 ;;;;
 ;;;; Author: Robert Smith
 
-(in-package #:cl-quil)
+(in-package #:cl-quil.frontend)
 
 (define-transform expand-circuits (expand-circuits)
   "This transform expands all circuits to create a circuit-free program.")