--- a/src/analysis/qubits-needed.lisp
+++ b/src/analysis/qubits-needed.lisp
@@ -2,7 +2,7 @@
 ;;;;
 ;;;; Author: Robert Smith
 
-(in-package #:cl-quil)
+(in-package #:cl-quil.frontend)
 
 ;;; This analysis counts the number of qubits needed to run a program.