--- a/src/classical-memory.lisp
+++ b/src/classical-memory.lisp
@@ -2,7 +2,7 @@
 ;;;;
 ;;;; Author: Robert Smith
 
-(in-package #:cl-quil)
+(in-package #:cl-quil.frontend)
 
 (define-condition quil-memory-model-error (a:simple-parse-error)
   ()