This factors out a reasonable subset of `cl-quil` into `cl-quil/frontend`. In particular, the frontend includes
- AST and parsing code
- clifford stuff
- analysis routines & transforms
- some oddball miscellaneous stuff (utilities, magicl constructors, etc) needed by the rest of the frontend

We also introduce the `cl-quil.frontend` package, exporting the symbols needed by `cl-quil` and `qvm`. 
This addresses #758 and (and along the way, #760).

Finally, we update some of the compressor routines to use the `unitary-qvm` for matrix computations.

This is a WIP in part because I am still fussing with the `cl-quil.frontend` package.

To get a sense of the the performance implications, I ran the benchmark code written by @mhdavid-hrl. Here's what I see on the current master (3ec4629) vs this branch (only looking at a few smaller chips)

`(on master)`
```
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| STATIC/FULLY-CONNECTED 10  |     2.20 |     0 |        2 |
| STATIC/FULLY-CONNECTED 15  |     3.89 |     0 |        2 |
| STATIC/FULLY-CONNECTED 20  |     6.59 |     0 |        2 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| BELL/FULLY-CONNECTED 10    |     3.90 |     0 |        9 |
| BELL/FULLY-CONNECTED 15    |     7.27 |     0 |       14 |
| BELL/FULLY-CONNECTED 20    |    11.75 |     0 |       19 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| QFT/FULLY-CONNECTED 10     |     8.64 |     0 |       37 |
| QFT/FULLY-CONNECTED 15     |    19.08 |     0 |       57 |
| QFT/FULLY-CONNECTED 20     |    35.08 |     0 |       77 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| HADAMARD/FULLY-CONNECTED 10 |     1.70 |     0 |        0 |
| HADAMARD/FULLY-CONNECTED 15 |     3.60 |     0 |        0 |
| HADAMARD/FULLY-CONNECTED 20 |     6.34 |     0 |        0 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| STATIC/LINEAR 10           |     0.78 |     0 |        2 |
| STATIC/LINEAR 15           |     1.02 |     0 |        2 |
| STATIC/LINEAR 20           |     1.34 |     0 |        2 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| BELL/LINEAR 10             |     2.81 |     0 |        9 |
| BELL/LINEAR 15             |     3.88 |     0 |       14 |
| BELL/LINEAR 20             |     6.12 |     0 |       19 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| QFT/LINEAR 10              |     5.44 |    61 |       79 |
| QFT/LINEAR 15              |     4.22 |   148 |      127 |
| QFT/LINEAR 20              |     7.79 |   276 |      175 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| HADAMARD/LINEAR 10         |     0.08 |     0 |        0 |
| HADAMARD/LINEAR 15         |     0.13 |     0 |        0 |
| HADAMARD/LINEAR 20         |     0.17 |     0 |        0 |
+----------------------------+----------+-------+----------+
```

`(this branch)`
```
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| STATIC/FULLY-CONNECTED 10  |     0.28 |     0 |        2 |
| STATIC/FULLY-CONNECTED 15  |     0.54 |     0 |        2 |
| STATIC/FULLY-CONNECTED 20  |     0.90 |     0 |        2 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| BELL/FULLY-CONNECTED 10    |     0.65 |     0 |        9 |
| BELL/FULLY-CONNECTED 15    |     1.13 |     0 |       14 |
| BELL/FULLY-CONNECTED 20    |     1.74 |     0 |       19 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| QFT/FULLY-CONNECTED 10     |     1.65 |     0 |       37 |
| QFT/FULLY-CONNECTED 15     |     3.82 |     0 |       57 |
| QFT/FULLY-CONNECTED 20     |     7.73 |     0 |       77 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| HADAMARD/FULLY-CONNECTED 10 |     0.25 |     0 |        0 |
| HADAMARD/FULLY-CONNECTED 15 |     0.53 |     0 |        0 |
| HADAMARD/FULLY-CONNECTED 20 |     0.93 |     0 |        0 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| STATIC/LINEAR 10           |     0.12 |     0 |        2 |
| STATIC/LINEAR 15           |     0.15 |     0 |        2 |
| STATIC/LINEAR 20           |     0.17 |     0 |        2 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| BELL/LINEAR 10             |     0.49 |     0 |        9 |
| BELL/LINEAR 15             |     0.88 |     0 |       14 |
| BELL/LINEAR 20             |     1.18 |     0 |       19 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| QFT/LINEAR 10              |     1.88 |    61 |       79 |
| QFT/LINEAR 15              |     4.26 |   150 |      127 |
| QFT/LINEAR 20              |     7.68 |   276 |      175 |
+----------------------------+----------+-------+----------+
+----------------------------+----------+-------+----------+
|        NAME                | TIME (s) | SWAPS | 2Q DEPTH |
+----------------------------+----------+-------+----------+
| HADAMARD/LINEAR 10         |     0.08 |     0 |        0 |
| HADAMARD/LINEAR 15         |     0.12 |     0 |        0 |
| HADAMARD/LINEAR 20         |     0.17 |     0 |        0 |
+----------------------------+----------+-------+----------+
```

It's a significant speedup, and I'm scratching my head wondering whether I've goofed anything or if these numbers are real.