@kilimanjaro 2021-11-04
Testing
- `cl-quil` passes
- `cl-quil/quilec` passes
- `cl-quil/quilt` passes
- `cl-quil-benchmarking` passes
- `quilc` passes

On the QVM side (cf https://github.com/quil-lang/qvm/pull/283)
- `qvm` passes
- `qvm-app` passes
- `qvm-benchmarks` passes
- `qvm-app-ng` has a few RPCQ errors which seem to be unrelated to this PR, but the rest passes
- `dqvm2` TBD (`brew install open-mpi` will be done within a few days...)