@karlosz 2021-11-11
<img width="1792" alt="Screen Shot 2021-11-11 at 12 28 23" src="https://user-images.githubusercontent.com/6234337/141366014-9ffa3804-0fc6-46ae-9942-3cea0a4781ae.png">

This flamegraph explains some of the speedup. Apparently `COMPILE` is called from `make-matrix-from-gate`, and those calls take up about 90% of compile time.